refactor: move DatanodeStat related structs to common_meta
