+++ src/meta-srv/src/handler/region_lease_handler.rs
@@ -103,6 +103,7 @@ mod test {
     use std::collections::{HashMap, HashSet};
     use std::sync::Arc;
 
+    use common_meta::datanode::{RegionStat, Stat};
     use common_meta::distributed_time_constants;
     use common_meta::key::table_route::TableRouteValue;
     use common_meta::key::test_utils::new_test_table_info;
@@ -115,7 +116,6 @@ mod test {
     use store_api::storage::RegionId;
 
     use super::*;
-    use crate::handler::node_stat::{RegionStat, Stat};
     use crate::metasrv::builder::MetasrvBuilder;
 
     fn new_test_keeper() -> RegionLeaseKeeper {
