+++ src/meta-srv/src/key/datanode.rs
@@ -14,6 +14,7 @@
 
 use std::str::FromStr;
 
+use common_meta::datanode::DatanodeStatKey;
 use common_meta::ClusterId;
 use lazy_static::lazy_static;
 use regex::Regex;
@@ -22,7 +23,6 @@ use snafu::{ensure, OptionExt, ResultExt};
 
 use crate::error;
 use crate::error::Result;
-use crate::handler::node_stat::Stat;
 
 pub(crate) const DATANODE_LEASE_PREFIX: &str = "__meta_datanode_lease";
 const INACTIVE_REGION_PREFIX: &str = "__meta_inactive_region";
@@ -52,118 +52,12 @@ impl DatanodeLeaseKey {
     }
 }
 
-#[derive(Debug, Clone, Copy, Eq, PartialEq, Hash)]
-pub struct DatanodeStatKey {
-    pub cluster_id: ClusterId,
-    pub node_id: u64,
-}
-
-impl DatanodeStatKey {
-    pub fn prefix_key() -> Vec<u8> {
-        format!("{DATANODE_STAT_PREFIX}-").into_bytes()
-    }
-}
-
 impl From<&DatanodeLeaseKey> for DatanodeStatKey {
     fn from(lease_key: &DatanodeLeaseKey) -> Self {
         DatanodeStatKey {
             cluster_id: lease_key.cluster_id,
             node_id: lease_key.node_id,
         }
-    }
-}
-
-impl From<DatanodeStatKey> for Vec<u8> {
-    fn from(value: DatanodeStatKey) -> Self {
-        format!(
-            "{}-{}-{}",
-            DATANODE_STAT_PREFIX, value.cluster_id, value.node_id
-        )
-        .into_bytes()
-    }
-}
-
-impl FromStr for DatanodeStatKey {
-    type Err = error::Error;
-
-    fn from_str(key: &str) -> Result<Self> {
-        let caps = DATANODE_STAT_KEY_PATTERN
-            .captures(key)
-            .context(error::InvalidStatKeySnafu { key })?;
-
-        ensure!(caps.len() == 3, error::InvalidStatKeySnafu { key });
-
-        let cluster_id = caps[1].to_string();
-        let node_id = caps[2].to_string();
-        let cluster_id: u64 = cluster_id.parse().context(error::ParseNumSnafu {
-            err_msg: format!("invalid cluster_id: {cluster_id}"),
-        })?;
-        let node_id: u64 = node_id.parse().context(error::ParseNumSnafu {
-            err_msg: format!("invalid node_id: {node_id}"),
-        })?;
-
-        Ok(Self {
-            cluster_id,
-            node_id,
-        })
-    }
-}
-
-impl TryFrom<Vec<u8>> for DatanodeStatKey {
-    type Error = error::Error;
-
-    fn try_from(bytes: Vec<u8>) -> Result<Self> {
-        String::from_utf8(bytes)
-            .context(error::StatKeyFromUtf8Snafu {})
-            .map(|x| x.parse())?
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(transparent)]
-pub struct DatanodeStatValue {
-    pub stats: Vec<Stat>,
-}
-
-impl DatanodeStatValue {
-    /// Get the latest number of regions.
-    pub fn region_num(&self) -> Option<u64> {
-        self.stats.last().map(|x| x.region_num)
-    }
-
-    /// Get the latest node addr.
-    pub fn node_addr(&self) -> Option<String> {
-        self.stats.last().map(|x| x.addr.clone())
-    }
-}
-
-impl TryFrom<DatanodeStatValue> for Vec<u8> {
-    type Error = error::Error;
-
-    fn try_from(stats: DatanodeStatValue) -> Result<Self> {
-        Ok(serde_json::to_string(&stats)
-            .context(error::SerializeToJsonSnafu {
-                input: format!("{stats:?}"),
-            })?
-            .into_bytes())
-    }
-}
-
-impl FromStr for DatanodeStatValue {
-    type Err = error::Error;
-
-    fn from_str(value: &str) -> Result<Self> {
-        serde_json::from_str(value).context(error::DeserializeFromJsonSnafu { input: value })
-    }
-}
-
-impl TryFrom<Vec<u8>> for DatanodeStatValue {
-    type Error = error::Error;
-
-    fn try_from(value: Vec<u8>) -> Result<Self> {
-        String::from_utf8(value)
-            .context(error::StatValueFromUtf8Snafu {})
-            .map(|x| x.parse())?
     }
 }
 
@@ -254,29 +148,6 @@ mod tests {
     }
 
     #[test]
-    fn test_stat_val_round_trip() {
-        let stat = Stat {
-            cluster_id: 0,
-            id: 101,
-            region_num: 100,
-            ..Default::default()
-        };
-
-        let stat_val = DatanodeStatValue { stats: vec![stat] };
-
-        let bytes: Vec<u8> = stat_val.try_into().unwrap();
-        let stat_val: DatanodeStatValue = bytes.try_into().unwrap();
-        let stats = stat_val.stats;
-
-        assert_eq!(1, stats.len());
-
-        let stat = stats.first().unwrap();
-        assert_eq!(0, stat.cluster_id);
-        assert_eq!(101, stat.id);
-        assert_eq!(100, stat.region_num);
-    }
-
-    #[test]
     fn test_lease_key_round_trip() {
         let key = DatanodeLeaseKey {
             cluster_id: 0,
@@ -287,67 +158,6 @@ mod tests {
         let new_key: DatanodeLeaseKey = key_bytes.try_into().unwrap();
 
         assert_eq!(new_key, key);
-    }
-
-    #[test]
-    fn test_get_addr_from_stat_val() {
-        let empty = DatanodeStatValue { stats: vec![] };
-        let addr = empty.node_addr();
-        assert!(addr.is_none());
-
-        let stat_val = DatanodeStatValue {
-            stats: vec![
-                Stat {
-                    addr: "1".to_string(),
-                    ..Default::default()
-                },
-                Stat {
-                    addr: "2".to_string(),
-                    ..Default::default()
-                },
-                Stat {
-                    addr: "3".to_string(),
-                    ..Default::default()
-                },
-            ],
-        };
-        let addr = stat_val.node_addr().unwrap();
-        assert_eq!("3", addr);
-    }
-
-    #[test]
-    fn test_get_region_num_from_stat_val() {
-        let empty = DatanodeStatValue { stats: vec![] };
-        let region_num = empty.region_num();
-        assert!(region_num.is_none());
-
-        let wrong = DatanodeStatValue {
-            stats: vec![Stat {
-                region_num: 0,
-                ..Default::default()
-            }],
-        };
-        let right = wrong.region_num();
-        assert_eq!(Some(0), right);
-
-        let stat_val = DatanodeStatValue {
-            stats: vec![
-                Stat {
-                    region_num: 1,
-                    ..Default::default()
-                },
-                Stat {
-                    region_num: 0,
-                    ..Default::default()
-                },
-                Stat {
-                    region_num: 2,
-                    ..Default::default()
-                },
-            ],
-        };
-        let region_num = stat_val.region_num().unwrap();
-        assert_eq!(2, region_num);
     }
 
     #[test]
